total decimal

+++ VolcomInventory/FormAdjustmentOGDet.Designer.vb
@@ -534,7 +534,7 @@ Partial Class FormAdjustmentOGDet
         Me.GridColumn16.OptionsColumn.ReadOnly = True
         Me.GridColumn16.Summary.AddRange(New DevExpress.XtraGrid.GridSummaryItem() {New DevExpress.XtraGrid.GridColumnSummaryItem(DevExpress.Data.SummaryItemType.Sum, "GridColumn16", "{0:N2}")})
         Me.GridColumn16.UnboundExpression = "[qty] * [value]"
-        Me.GridColumn16.UnboundType = DevExpress.Data.UnboundColumnType.[Integer]
+        Me.GridColumn16.UnboundType = DevExpress.Data.UnboundColumnType.[Decimal]
         Me.GridColumn16.Visible = True
         Me.GridColumn16.VisibleIndex = 7
         '
